flydra2: stream some files without gz compression
For whatever reason, drop() is not called on WritingState when a panic
happens. (Ideally this should be fixed.) To nevertheless maximize
chances that the textlog and experiment_info data are saved to disk, we
save them without .gz compression. Combined with flushing the CSV
writer, this does indeed get things written to disk.