strand-cam: restore checkerboard cal to browser UI
This fixes an issue inadvertently introduced in
5e4d255d07fafdf7139046fe1e414fffbcc2f4d1 in which the browser UI was not
compiled with, and did not show, the checkerboard calibration
components.

This additionally introduces runtime checks so that a similar issue will
be caught again as the browser UI code will raise an error if it
attempts to deserialize data with additional, unexpected fields. This
would have caught the issue sooner.