py-strandcam: remove unused Python wrapper
This was not documented and as far as I know,
never used "for real". It was not trivial to keep
it updated with various changes to the python
distutils/setuptools/etc ecosystem and probably if
done again, I would do it differently by using
pyo3 instead of milksnake and cffi.